Rename WindowManager classes and move them to WindowContainer folder
