+++ src/Overlay/WindowContainer/WindowContainerImpl.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "WindowContainer.h"
+
+class WindowContainerImpl : public WindowContainer
+{
+public:
+	~WindowContainerImpl() override = default;
+protected:
+	void FillWindowContainer() override;
+};
