+++ src/Overlay/WindowContainer/WindowType.h
@@ -1,0 +1,12 @@
+#pragma once
+
+enum WindowType_
+{
+	WindowType_Main,
+	WindowType_Debug,
+	WindowType_Log,
+	WindowType_Donators,
+	WindowType_UpdateNotifier,
+	WindowType_PaletteEditor,
+	WindowType_CustomHud
+};
